+++ entrega 3 grupo 2/entrega 3 grupo 2 solo diccionario NO mapa/Product.cs
@@ -18,8 +18,8 @@ namespace entrega_01_grupo_2
         protected double DiseaseProbability;
         protected int DiseasePenalty;
 
-        public Product(string name, double purchase_price, double base_selling_price, double water_use_per_turn, int min_water_lv, 
-            int water_penalty, int production_time, int disease_probability, int disease_penalty)
+        public Product(string name, double purchase_price, double base_selling_price, double water_use_per_turn, double min_water_lv, 
+            double water_penalty, double production_time, int disease_probability, double disease_penalty)
         {
             this.Name = name;
             this.PurchasePrice = purchase_price;
