+++ entrega 3 grupo 2/entrega 3 grupo 2 solo diccionario NO mapa/Seed.cs
@@ -18,7 +18,7 @@ namespace entrega_01_grupo_2
         private int WeedPenalty;
 
         public Seed(string Name, double PurchasePrice, double BaseSellingPrice, double WaterConsumptionPerTurn, int MinWaterLv, 
-            int WaterPenalty, int ProductionTime, int DeseaseProbability, int DeseasePenalty, int price_variation,
+            int WaterPenalty, int ProductionTime, int DeseaseProbability, int DeseasePenalty, double price_variation,
             int nutrient_use_per_turn, int min_nutrient_lv, int nutrient_penalty, int worm_probability, int worm_penalty, 
             int weed_probability, int weed_penalty) : 
             base(Name, PurchasePrice, BaseSellingPrice, WaterConsumptionPerTurn,
